[BR-104] : LLM + 동적/정적 장애물 구분 명령 함수 구현 80%
