--- alba_gpt/AlbaGPT_LLM/.gitignore
@@ -1,1 +1,0 @@
-.env
