--- alba_gpt/AlbaGPT_LLM/contents/response/CAMREA_ON_response.json
@@ -1,18 +1,0 @@
-{
-    "pinky_id": "1",
-    "pinky_task": "CAMREA_ON",
-    "pinky_response": "이미지에는 여러 개의 상자가 보입니다. 각 상자에는 음식 이미지가 있으며, 'stock' 및 'sold'라는 레이블이 붙어 있습니다. 이는 재고 여부를 나타내는 것으로 보입니다."
-}
-
-{
-    "pinky_id": "1",
-    "pinky_task": "CAMREA_ON",
-    "pinky_response": "이미지에는 팔과 손이 보이며, 검은색 옷을 입고 있는 것 같습니다. 배경에는 책상과 물건들이 보입니다."
-}
-
-{
-    "pinky_id": "1",
-    "pinky_task": "CAMREA_ON",
-    "pinky_response": "사진에는 책상 위에 시계와 물병이 놓여 있으며, 한 사람이 앉아 있습니다. 배경에는 문과 스크린이 보입니다."
-}
-
