--- alba_gpt/AlbaGPT_LLM/contents/response/EMERGENCY_response.json
@@ -1,35 +1,0 @@
-{
-    "pinky_id": "",
-    "pinky_task": "EMERGENCY",
-    "pinky_response": "지금 매장에 도둑이 들었습니다! 즉시 대피하세요! 🚨",
-    "table_id": ""
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "EMERGENCY",
-    "pinky_response": "핑키들, 지금 매장에 도둑이 들었어! 즉시 경계 태세를 취하겠습니다! 🚨",
-    "table_id": ""
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "EMERGENCY",
-    "pinky_response": "도둑이 들어왔습니다! 즉시 경비를 호출하고 손님들을 안전하게 대피시키겠습니다!",
-    "table_id": ""
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "EMERGENCY",
-    "pinky_response": "도둑이 들어왔습니다! 즉시 경고하고 대피하세요! 🚨",
-    "table_id": ""
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "EMERGENCY",
-    "pinky_response": "도둑이 들어왔습니다! 즉시 경고하고 안전하게 대피하세요! 🚨",
-    "table_id": ""
-}
-
