--- alba_gpt/AlbaGPT_LLM/contents/response/GREETINGS_response.json
@@ -1,78 +1,0 @@
-{
-    "pinky_id": "",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "안녕하세요 사장님! 좋은 하루 되세요 😊"
-}
-
-{
-    "pinky_id": "1",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "오늘 기분이 아주 좋습니다 사장님! 함께 즐거운 하루 만들어봐요 😊"
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "안녕하세요 사장님! 좋은 하루 되세요 😊"
-}
-
-{
-    "pinky_id": "1",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "오늘 기분이 아주 좋습니다 사장님! 신나는 하루예요 🎶"
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "안녕하세요 사장님! 좋은 하루 되세요 😊"
-}
-
-{
-    "pinky_id": "1",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "오늘 기분이 좋습니다 사장님! 즐거운 하루 되세요 😊"
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "안녕하세요 사장님! 좋은 하루 되세요 😊"
-}
-
-{
-    "pinky_id": "1",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "오늘 기분이 아주 좋습니다 사장님! 신나는 하루예요 🎶"
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "안녕하세요 사장님! 좋은 하루 되세요 😊"
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "안녕하세요 사장님! 좋은 하루 되세요 😊"
-}
-
-{
-    "pinky_id": "1",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "오늘 기분이 아주 좋습니다 사장님! 신나는 하루예요 🎶"
-}
-
-{
-    "pinky_id": "",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "안녕하세요 사장님! 좋은 하루 되세요 😊"
-}
-
-{
-    "pinky_id": "1",
-    "pinky_task": "GREETINGS",
-    "pinky_response": "오늘 기분이 아주 좋습니다 사장님! 신나는 하루예요 🎶"
-}
-
