--- alba_gpt/AlbaGPT_LLM/contents/response/SERVING_response.json
@@ -1,21 +1,0 @@
-{
-    "pinky_id": "2",
-    "pinky_task": "SERVING",
-    "pinky_response": "5번 테이블로 신선한 셀러드를 서빙하겠습니다 🥗",
-    "table_id": "5번 테이블"
-}
-
-{
-    "pinky_id": "2",
-    "pinky_task": "SERVING",
-    "pinky_response": "5번 테이블에 신선한 셀러드를 서빙하겠습니다 🥗",
-    "table_id": "5번 테이블"
-}
-
-{
-    "pinky_id": "2",
-    "pinky_task": "SERVING",
-    "pinky_response": "네! 5번 테이블에 셀러드 하나를 서빙하겠습니다 🥗",
-    "table_id": "5번 테이블"
-}
-
