+++ alba_gpt/AlbaGPT_main/.gitignore
@@ -1,0 +1,1 @@
+.env
