+++ alba_gpt/AlbaGPT_main/contents/example/obstacle_example.csv
@@ -1,0 +1,41 @@
+response
+가 길을 가로막고 있어요 🥲
+가 막고 있어요! 저를 도와주세요! 🥹
+가 제 앞을 가로막고 있어서 지나갈 수 없어요 😣
+앞에 가 있어서 조금 곤란해요 😅
+이 앞에 가 있어서 다른 길로 돌아가야 할 것 같아요 😕
+가 제 경로에 있어서 진행이 어려워요 😓
+저기 가 보여요… 조심히 피해갈게요! 🚧
+가 있어서 잠깐 멈춰 있을게요 🙏
+가 막고 있어서 지나가지 못해요 😢
+가 있어요! 이쪽으로 도와주실 수 있을까요? 🙋‍♀️
+가 있어서 조심히 피해야 할 것 같아요 🫣
+조심하세요! 가 길 한가운데 있어요 😬
+가 있어서 앞으로 못 가고 있어요 🥺
+가 있어서 제가 멈춰섰어요 😔
+가 앞에 있어서 움직일 수가 없어요 😥
+경로에 가 있습니다. 안전을 위해 정지할게요 🚨
+앞을 보니 가 있네요. 잠시 기다릴게요! ⏸️
+가 저를 막고 있어서 난감해요 😵
+가 보여요! 조심히 이동할게요 🚶
+가 앞으로 와서 움직일 수 없어요 😟
+누가 가를 이곳에 뒀을까요? 지나가기 어렵네요 🧐
+가 방해가 돼서 멈췄어요 😔
+죄송해요, 가 있어서 우회가 필요해요 🙇‍♂️
+가 앞을 막고 있어서 대기 중이에요 ⏳
+이 앞에 가 보이는데요, 어떻게 할까요? 🤔
+가 앞에 있어서 제가 가로막혔어요 🧍
+지금 가 때문에 멈춰 있어요 🙁
+가 앞에서 버티고 있어요! 😤
+앞에 가 있어서 주의가 필요해요 ⚠️
+가 제 경로를 완전히 막고 있어요 🚫
+가 앞으로 다가와서 무서워요 🥶
+가 있어서 피해야 해요! 😱
+어이쿠, 가 있어서 멈췄어요 🫢
+가 보여서 멈추는 중이에요! 🛑
+가 앞으로 와서 저는 움직일 수 없어요 😓
+가 앞에 있어서 길이 막혔어요 🤷
+가 있어서 진로 확보가 어렵습니다 📡
+저기 가 보여요… 누가 좀 옮겨주실 수 있을까요? 🙋
+가 경로를 막고 있어서 저는 잠시 멈춰 있을게요 😌
+가 있기 때문에 제가 앞으로 나아갈 수 없어요 😞
