+++ alba_gpt/AlbaGPT_main/contents/response/MAINTENANCE_response.json
@@ -1,0 +1,21 @@
+{
+    "pinky_id": "3",
+    "pinky_task": "MAINTENANCE",
+    "pinky_response": "충전이 필요하다는 신호를 받았어요. 이동합니다! 🚧",
+    "table_id": ""
+}
+
+{
+    "pinky_id": "3",
+    "pinky_task": "MAINTENANCE",
+    "pinky_response": "충전이 필요하다는 신호를 받았어요. 이동합니다! 🚧",
+    "table_id": ""
+}
+
+{
+    "pinky_id": "3",
+    "pinky_task": "MAINTENANCE",
+    "pinky_response": "충전이 필요하다는 신호를 받았어요. 이동합니다! 🚧",
+    "table_id": ""
+}
+
